pyproject.toml
[project]
name = "locatel-demo-backend"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "django>=5.1.3",
    "djangorestframework>=3.15.2",
    "markdown>=3.7",
    "django-filter>=24.3",
    "django-cors-headers>=4.6.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/init"]
